+++ display/help.go
@@ -17,7 +17,7 @@ b, Alt-Left|Go back a page
 f, Alt-Right|Go forward a page
 g|Go to top of document
 G|Go to bottom of document
-spacebar|Open bar at the bottom - type a URL or link number
+spacebar|Open bar at the bottom - type a URL, link number, or search term. You can also type two dots (..) to go up a directory in the URL.
 Enter|On a page this will start link highlighting. Press Tab and Shift-Tab to pick different links. Press enter again to go to one.
 Shift-NUMBER|Go to a specific tab.
 Shift-0, )|Go to the last tab.