+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Alt-Left and Alt-Right for history navigation (#23)
+- You can type `..` in the bottom bar to go up a directory in the URL (#21)
 
 ### Changed
 - Bottom bar now says `URL/Num./Search: ` when space is pressed